+++ www/modules/email/LinkAttachmentDialog.js
@@ -2,6 +2,8 @@ GO.email.LinkAttachmentDialog = Ext.extend(go.links.CreateLinkWindow,{
 
 	attachmentItem : null, // If this is set to null, then it saves all attachments of the message.
 	messagePanel : null,
+	isEmailEditor : false,
+	attachmentsView : null,
 	
 	constructor : function(config){
 		
@@ -15,7 +17,12 @@ GO.email.LinkAttachmentDialog = Ext.extend(go.links.CreateLinkWindow,{
 
 		GO.email.LinkAttachmentDialog.superclass.constructor.call(this,config);
 	},
-	
+	setEmailEditor : function(emailEditor) {
+		this.isEmailEditor = emailEditor;
+	},
+	setAttachmentsView : function(attachmentsView) {
+		this.attachmentsView = attachmentsView;
+	},
 	link : function()	{
 		var record = this.grid.getSelectionModel().getSelected();
 		
@@ -41,7 +48,28 @@ GO.email.LinkAttachmentDialog = Ext.extend(go.links.CreateLinkWindow,{
 			success:function(response, options, result){
 				
 				if(GO.util.empty(this.attachmentItem)){
-					this.saveAllToItem(record, result.files_folder_id);
+					if(this.isEmailEditor) {
+						GO.files.createSelectFileBrowser();
+
+						GO.selectFileBrowser.setFileClickHandler(function(){
+
+							var paths = [];
+							var selections = GO.selectFileBrowser.getSelectedGridRecords();
+							for (var i = 0; i < selections.length; i++)
+								paths.push(selections[i].data.path);
+
+							this.attachmentsView.afterUpload({addFileStorageFiles:Ext.encode(paths)});
+							GO.selectFileBrowserWindow.hide();
+						}, this);
+
+						GO.selectFileBrowser.setFilesFilter('');
+						GO.selectFileBrowser.setRootID(result.files_folder_id,result.files_folder_id);
+						GO.selectFileBrowserWindow.show();
+						this.hide();
+					} else {
+						this.saveAllToItem(record, result.files_folder_id);
+					}
+					
 				} else {
 					this.saveToItem(record, result.files_folder_id);
 				}
@@ -57,7 +85,6 @@ GO.email.LinkAttachmentDialog = Ext.extend(go.links.CreateLinkWindow,{
 	},
 
 	saveToItem : function(record,files_folder_id){
-
 		if(!GO.files.saveAsDialog){
 			GO.files.saveAsDialog = new GO.files.SaveAsDialog();
 		}